move logs to wow directory
